requirements.txt
Flask==2.1.2
requests==2.26.0
mysql-connector-python==8.0.28
PyYAML==6.0
